feat(analysis): extract mimo functions to separate class (#3742)
