+++ analysis/filter/MergeFilter/src/main/kotlin/de/maibornwolff/codecharta/filter/mergefilter/mimo/Mimo.kt
@@ -1,0 +1,84 @@
+package de.maibornwolff.codecharta.filter.mergefilter.mimo
+
+import de.maibornwolff.codecharta.filter.mergefilter.ParserDialog
+import de.maibornwolff.codecharta.util.Logger
+import java.io.File
+
+class Mimo {
+    companion object {
+        fun generateProjectGroups(sourceFiles: List<File>, maxLevenshteinDistance: Int): List<Pair<Boolean, List<File>>> {
+            val mutableSourceFiles: MutableList<File> = sourceFiles.toMutableList()
+            val groupedFiles: MutableList<Pair<Boolean, List<File>>> = mutableListOf()
+            while (mutableSourceFiles.isNotEmpty()) {
+                val currentFile = mutableSourceFiles.removeFirst()
+                var exactMatch = true
+                val currentFileList: MutableList<File> = mutableListOf()
+
+                mutableSourceFiles.forEach {
+                    val prefixMatch = evaluatePrefixMatch(currentFile, it, maxLevenshteinDistance)
+                    if (prefixMatch == 0) {
+                        currentFileList.add(it)
+                    } else if (prefixMatch == 1) {
+                        exactMatch = false
+                        currentFileList.add(it)
+                    }
+                }
+
+                mutableSourceFiles.removeAll(currentFileList)
+                currentFileList.add(currentFile)
+
+                if (currentFileList.size > 1) {
+                    groupedFiles.add(Pair(exactMatch, currentFileList))
+                } else {
+                    Logger.debug { "Discarded ${currentFile.name} as a potential group" }
+                }
+            }
+            return groupedFiles
+        }
+
+        private fun evaluatePrefixMatch(original: File, comparison: File, maxLevenshteinDistance: Int): Int {
+            val ogPrefix = original.name.substringBefore(".")
+            val compPrefix = comparison.name.substringBefore(".")
+            return if (ogPrefix == compPrefix) {
+                0
+            } else if (maxLevenshteinDistance > 0 && levenshteinDistance(ogPrefix, compPrefix) <= maxLevenshteinDistance) {
+                1
+            } else {
+                -1
+            }
+        }
+
+        fun retrieveGroupName(files: List<File>): String {
+            val filePrefixes = files.map { it.name.substringBefore(".") }.toSet()
+            if (filePrefixes.size == 1) return filePrefixes.first()
+            return ParserDialog.askForMimoPrefix(filePrefixes)
+        }
+
+        private fun levenshteinDistance(lhs: CharSequence, rhs: CharSequence): Int {
+            val lhsLength = lhs.length
+            val rhsLength = rhs.length
+
+            var cost = IntArray(rhsLength + 1) { it }
+            var newCost = IntArray(rhsLength + 1)
+
+            for (i in 1..lhsLength) {
+                newCost[0] = i
+
+                for (j in 1..rhsLength) {
+                    val match = if (lhs[i - 1] == rhs[j - 1]) 0 else 1
+                    val costReplace = cost[j - 1] + match
+                    val costInsert = cost[j] + 1
+                    val costDelete = newCost[j - 1] + 1
+
+                    newCost[j] = minOf(costInsert, costDelete, costReplace)
+                }
+
+                val swap = cost
+                cost = newCost
+                newCost = swap
+            }
+
+            return cost[rhsLength]
+        }
+    }
+}
